Use @import JSDoc tags for TypeScript >= 5.5
These are easier to read than the /** @typedef */ workaround, and also
take up less space. For each import, a separate @typedef is required.

TypeScript supports /** @import */ since version 5.5, see
https://www.typescriptlang.org/docs/handbook/release-notes/typescript-5-5.html#the-jsdoc-import-tag

The script detects the current TypeScript version and uses @typedef for version < 5.5,
and @import for versions >= 5.5. Perhaps you could also add an option for which import
style to use, but I prefer less options. transpileFile has an optiona; `tsVersion` argument,
which is needed for tests and could also be used by consumers if needed.